mkimg: make MBR and EBR honor alignment, and change default alignment to 4k
Summary:
mkimg previously, even if you specified -P 4k, would not align
MBR images to a 4k boundary.  This hasn't been necessary for MBR
for almost 20 years.

As 4k drives are very common, change mkimg to default to a 4k
alignment.  I would be fine w/ uping this to 16k or 64k.
This gives images the benefit w/o having to track down all the
various scripts that make an image to fix them.

Test Plan:
If you see the changes to the test program, it mounts and outputs
the gpart show info for the raw images.  I used this to verify
that things are now properly aligned...

Reviewers: imp, gjb, emaste, #manpages, marcel

Reviewed By: imp, marcel

Subscribers: greg_unrelenting.technology

Differential Revision: https://reviews.freebsd.org/D22068