inetd: a light introduction to capsicum
Summary:
This isn't a complete solution, but perhaps roughly the first half of one. This should effectively limit all fds opened by inetd, even those we hook up to child stdin/out/err, and enters capability mode for some of the builtin handlers.

I think the other half of the solution is to devise a syntax for inetd.conf to indicate that we should go ahead and enter capability mode for non-builtins.

Reviewers: markj, oshogbo

Subscribers: imp, jonathan

Differential Revision: https://reviews.freebsd.org/D23022