[PowerPC] Fix sensor attachment on PowerBook5,3
Summary:
After spending a while trying to figure out why the fans on my PowerBook G4 would always be at full speed, and the thermal sensors were missing, I realized that the ADT746x driver did not actually have support for the "Version 1" sensor format, and would only work on models that had extra nodes for the individual child sensors.

This patch changes the behavior for hwsensor-params-version == 1 to read everything out of the device properties.

Test Plan: Test on a variety of hardware and ensure this doesn't cause any regressions

Reviewers: #powerpc

Subscribers: jhibbits, imp, luporl, alfredo

Tags: #powerpc

Differential Revision: https://reviews.freebsd.org/D23809